+++ gardenproject/urls.py
@@ -3,11 +3,12 @@ from django.urls import include, path
 from rest_framework import routers
 from gardenapi.views import register_user, login_user
 from rest_framework.authtoken.views import obtain_auth_token
-from gardenapi.views import Posts, Topics
+from gardenapi.views import Posts, Topics, Comments
 
 router = routers.DefaultRouter(trailing_slash=False)
 router.register(r"posts", Posts, "post")
 router.register(r"topics", Topics, "topic")
+router.register(r"comments", Comments, "comment")
 
 
 urlpatterns = [
